change le type de la variable de filtration de int vers double et ajoute une convertion implicite vers msg.data.

+++ include/WMLiftingColumnHardwareInterface.h
@@ -55,7 +55,7 @@ namespace wm_lifting_column_hardware_interface
         int mResolution;
         double posBuffer;
         int mBumperTimeCounter;
-        int mFilteredCmd;
+        double mFilteredCmd;
     };
 }
 #endif //PROJECT_WMLiftingColumnHardwareInterface_H
